chore(gradle): Update dependencies and Java version
Updated several library dependencies and the targeted Java version in the build.gradle.kts file. The changes provide more features for testing and lay the groundwork for Kotlin toolchain with the latest LTS Java version. The paper's version was also updated to the latest snapshot for testing purposes, and Kotest and MockBukkit versions were introduced for holistic testing. Java version was also abstracted into variables.