chore(koin): Refactor Koin injection and setup
The injection method in the CraftContext class has been refactored to enhance the shared instance retrieving between plugins. A new injection method has been added to retrieve instances specific to a plugin id with a fallback option to the APIPlugin id.
Changes have also been made to cater to the non-nullability of the Koin instance in the _koins map, eliminating the option of retrieving a null instance.
The Koin setup on tests, APIPlugin and the plugin has been modified to reflect the changes in the injection methods and constants.